Rust rewrite
The previous version in C++ was brittle in may regards.
This change reimplements everything in Rust and makes error handling much better.

This further packages `es2command-sys` crate which provides a minimal set of headers from the epson source which allows to derive the epsonscan2 library interface without downloading the original source.